requirements.txt
Django==2.2.4
djangorestframework==3.10.2
isort==4.3.21
MarkupSafe==1.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
python-slugify==3.0.3
pytz==2019.2
sqlparse==0.3.0
text-unidecode==1.2
